packages/biometric/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
opencv-python==4.9.0.80
numpy==1.26.4
scikit-image==0.22.0
scikit-learn==1.4.1.post1
pillow==10.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
httpx==0.27.0
loguru==0.7.2
